requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
jinja2==3.1.4
sqlmodel==0.0.21
pydantic==2.8.2
python-multipart==0.0.9
starlette==0.38.5
colorama==0.4.6
PyPDF2==3.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4